refactor: data ディレクトリにデータ処理をまとめた
